+++ tests/Clock/SecondTest.php
@@ -1,0 +1,18 @@
+<?php
+declare(strict_types = 1);
+
+namespace Tests\Innmind\TimeContinuum\Clock;
+
+use Innmind\TimeContinuum\Clock\Second;
+use PHPUnit\Framework\TestCase;
+
+class SecondTest extends TestCase
+{
+    public function testInterface()
+    {
+        $second = Second::of(13);
+
+        $this->assertSame(13, $second->toInt());
+        $this->assertSame('13', $second->toString());
+    }
+}